fix(architecture): isolate private worker frames
