+++ src/jacobian/math/graphs/_independence_z3_worker.py
@@ -13,12 +13,12 @@ from jacobian._execution import (
     request_execution,
 )
 from jacobian._worker_errors import bind_worker_deadline, worker_execution_errors
+from jacobian._worker_protocol import encode_worker_result_frame
 from jacobian.math.graphs._independence_z3 import (
     _solve_independence_number_values_kernel,
 )
 from jacobian.math.graphs.independence import IndependenceNumberBudget
 from jacobian.math.graphs.values import SimpleUndirectedGraph
-from jacobian.process import encode_worker_result_frame
 
 
 def main() -> int:
