+++ src/jacobian/math/graphs/optimization/_chromatic_bipartition_worker.py
@@ -8,11 +8,11 @@ import time
 
 from jacobian._execution import request_execution
 from jacobian._worker_errors import bind_worker_deadline, worker_execution_errors
+from jacobian._worker_protocol import encode_worker_result_frame
 from jacobian.math.graphs.optimization._chromatic_bipartition import (
     ChromaticBipartitionRequest,
     _find_chromatic_bipartition_kernel,
 )
-from jacobian.process import encode_worker_result_frame
 
 
 def main() -> int:
