+++ src/jacobian/math/graphs/optimization/_chromatic_number_worker.py
@@ -13,13 +13,13 @@ from jacobian._execution import (
     request_execution,
 )
 from jacobian._worker_errors import bind_worker_deadline, worker_execution_errors
+from jacobian._worker_protocol import encode_worker_result_frame
 from jacobian.math.graphs.optimization._chromatic_number import (
     _search_chromatic_number_kernel,
 )
 from jacobian.math.graphs.optimization._coloring_models import (
     GraphChromaticNumberRequest,
 )
-from jacobian.process import encode_worker_result_frame
 
 
 def main() -> int:
