+++ tests/math/graphs/test_optimization_wall_budget.py
@@ -8,6 +8,9 @@ import pytest
 import z3
 
 from jacobian._execution import OperationExecutionTimeoutError
+from jacobian._worker_protocol import (
+    encode_worker_result_frame,
+)
 from jacobian.math.graphs import _independence_z3
 from jacobian.math.graphs.independence import (
     IndependenceNumberBudget,
@@ -32,7 +35,6 @@ from jacobian.math.graphs.values import SimpleUndirectedGraph
 from jacobian.process import (
     BoundedProcessResult,
     ProcessResourceLimits,
-    encode_worker_result_frame,
 )
